requirements.txt
Flask==2.0.3
flask-mysqldb==0.2.0
flask-bcrypt==0.7.1
python-dotenv==0.19.2